top-cards&decks

+++ src/main/java/com/example/demo/register/GetDeck.java
@@ -45,7 +45,7 @@ public class GetDeck {
 		
 		private Card commander;
 			
-		private int likeNumber; 
+		private Long likeNumber;
